backend/requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.37.0
sqlalchemy==2.0.43
pydantic==2.11.9
pydantic-settings==2.11.0
email-validator==2.2.0
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
pytesseract==0.3.13
Pillow==11.3.0
psycopg2-binary==2.9.10
